pyproject.toml
[project]
name = "instapilot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.1",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.42",
]
